Code cleanup
Also fixed the small mistake I did last time.

+++ source/items.h
@@ -340,7 +340,7 @@ public:
 	
 	void clear();
 
-	ItemType& operator[](int id) {return getItemType(id);}
+	ItemType& operator[](size_t id) {return getItemType(id);}
 	uint16_t getMaxID() const {return max_item_id;}
 
 	bool typeExists(int id) const;
